+++ src/charade/components/end-modal/end-modal.styled.ts
@@ -1,7 +1,15 @@
+import styled from 'styled-components/native'
+import { Modalize } from 'react-native-modalize'
 import { RESPONSIVE } from '@core/responsive/responsive'
 import { getThemeProp } from '@core/styled/theme'
 import { TEXT_SIZE } from '@core/text/text.constants'
-import styled from 'styled-components/native'
+
+export const CharadeEndModalize = styled(Modalize).attrs(props => ({
+  adjustToContentHeight: true,
+  modalStyle: {
+    backgroundColor: getThemeProp('backgroundPrimary')(props),
+  }
+}))``
 
 export const EndModalContainer = styled.View`
   height: ${RESPONSIVE.HEIGHT(50)}px;