+++ src/core/hooks/use-localize.hook.ts
@@ -1,18 +1,13 @@
 import * as React from 'react'
-import { NativeModules, Platform } from 'react-native'
 import { useSelector } from 'react-redux'
+import { SYSTEM_LANGUAGE } from '@core/system-language/system-language'
 import { LANGUAGE_CODES, Localization } from '../localize/localize.models'
 import { localize } from '../localize/localize'
 import { languageCodeSelector } from '../../settings/store/settings.selectors'
 
 export const useLocalize = (): () => Localization => {
-  const deviceLanguage = Platform.select({
-    ios: NativeModules?.SettingsManager?.settings?.AppleLocale,
-    android: NativeModules?.I18nManager?.localeIdentifier,
-  })
-
   const _languageCode = useSelector(languageCodeSelector)
-  const languageCode = _languageCode ?? deviceLanguage ?? LANGUAGE_CODES.EN
+  const languageCode = _languageCode ?? SYSTEM_LANGUAGE ?? LANGUAGE_CODES.EN
 
   const handleChange = React.useCallback(() => {
     return localize(languageCode)