+++ src/dashboard/dashboard.tsx
@@ -2,6 +2,7 @@ import * as React from 'react'
 import { Modalize } from 'react-native-modalize'
 import { Dimensions, ScrollView, View, Image } from 'react-native'
 import { useSelector } from 'react-redux'
+import { useNavigation } from '@react-navigation/native'
 import { LettersSlider } from '@core/letters-slider/letters-slider'
 import { useIsPremium } from '@core/hooks/use-is-premium.hook'
 import { LetterSliderDefaultValues } from '@core/letters-slider/models'
@@ -11,9 +12,11 @@ import { SelectedLetters, LettersGrid, DashboardButtonsAndModals } from './compo
 import { useSelectLetter, useSearchPossibleWords, useSoapModal, useDashboardRehydration } from './hooks'
 import { DashboardHost, DashboardSafeArea } from './dashboard.styled'
 import { nativeSearchEngineEnabledSelector } from '../settings/store/settings.selectors'
+import { SCREEN } from '../navigation/navigation.constants'
 
 export const Dashboard = () => {
   useDashboardRehydration()
+  const navigation = useNavigation<any>()
   const forceIndexModalizeRef = React.useRef<Modalize>(null)
   const forceIndexLetterIndexRef = React.useRef<null | number>(null)
   const nativeSearchEngineEnabled = useSelector(nativeSearchEngineEnabledSelector)
@@ -39,12 +42,21 @@ export const Dashboard = () => {
     }
   }
 
+  const handleSwitchToPlayground = () => {
+    navigation.navigate(SCREEN.DASHBOARD_PLAYGROUND)
+  }
+
+  const leftContentConfig = {
+    onPress: handleSwitchToPlayground,
+    icon: 'checkerboard',
+  }
+
   const ContentWrapper = Dimensions.get('screen').height < 800 ? ScrollView : View
 
   return React.useMemo(() => (
     <DashboardHost>
       <DashboardSafeArea>
-        <Header type="dashboard" />
+        <Header type="dashboard" leftContentConfig={leftContentConfig} />
         <SelectedLetters {...{ selectedLetters, onLongPressSelectedLetter, handleDeselectLetter }} />
 
         <ContentWrapper>