+++ src/mania/mania.tsx
@@ -2,6 +2,7 @@ import * as React from 'react'
 import WebView from 'react-native-webview'
 import { ActivityIndicator } from 'react-native'
 import { SafeAreaFlexContainer } from '@core/styled'
+import { Header } from '@core/header/header'
 
 export const Mania = () => {
   const MANIA_ENABLED = true
@@ -9,6 +10,7 @@ export const Mania = () => {
 
   return MANIA_ENABLED ? (
     <SafeAreaFlexContainer>
+      <Header type="more" title="SCRABBLEMANIA" backButton />
       <WebView
         source={{ uri }}
         originWhitelist={['*']}