+++ src/playground/playground.styled.ts
@@ -1,14 +1,24 @@
+import * as R from 'ramda'
 import styled, { ThemeProps } from 'styled-components/native'
-import { FlatListProps } from 'react-native'
-import { SafeAreaFlexContainer } from '@core/styled'
+import { Zoom } from 'react-native-reanimated-zoom'
+import { Host } from 'react-native-portalize'
+import { Dimensions, FlatListProps, StatusBar } from 'react-native'
 import { ThemeModel } from '@core/styled/models'
 import { getThemeProp } from '@core/styled/theme'
+import { PLAYGROUND_SPACING_MULTIPLIER } from './components/playground-field/playground-field.styled'
+import { BOTTOM_NAVIGATION_HEIGHT } from 'navigation/navigation.constants'
+import { isPlatform } from '@core/is-platform/is-platform'
 
-export const PlaygroundSafeArea = styled(SafeAreaFlexContainer).attrs(props => ({
-  justifyContent: 'space-between',
-  height: 100,
-  backgroundColor: getThemeProp('backgroundSecondary')(props),
-}))``
+export const PlaygroundHost = styled(Host)`
+  background-color: ${getThemeProp('backgroundPrimary')};
+`
+
+export const PlaygroundZoom = styled(Zoom).attrs({
+  maximumZoomScale: PLAYGROUND_SPACING_MULTIPLIER,
+})`
+  flex: 1;
+  z-index: 0;
+`
 
 export const PlaygroundFlatList = styled.FlatList.attrs({
   contentContainerStyle: {
@@ -19,9 +29,11 @@ export const PlaygroundFlatList = styled.FlatList.attrs({
   },
 })<FlatListProps<any> & any>``
 
-export const PlaygroundBottomContainer = styled.View<ThemeProps<ThemeModel>>`
-  width: 100%;
-  padding-top: 3px;
-  background-color: ${getThemeProp('backgroundSecondary')};
-  justify-content: center;
+interface PlaygroundBottomContainerProps {
+  bottomInset: number;
+}
+
+export const PlaygroundBottomContainer = styled.View<PlaygroundBottomContainerProps & ThemeProps<ThemeModel>>`
+  top: ${({ bottomInset }) => isPlatform('ios') ? bottomInset + 10 : BOTTOM_NAVIGATION_HEIGHT}px;
+  padding-top: 10px;
 `